net: bus: fix -EADDRINUSE reply for BUS_{ADD,DEL}_ADDR setsockopt
When calling BUS_ADD_ADDR and BUS_DEL_ADDR set socket options
the kernel replied with a -EADDRINUSE "Address already in use"
error even when the address is not used.

This is because the buffer given to setsockopt was expected by
kernel to be a struct bus_addr while the AF_BUS documentation
says that it has to be a struct sockaddr_bus.

Signed-off-by: Alban Crequy <alban.crequy@collabora.co.uk>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>