netfilter: nfdbus: check pointer sanity before use.
match_rule_search() can return a NULL pointer if it does not
find the required rule but the returned pointer was not being
checked for NULL in the calling function
bus_matchmaker_remove_rule_by_value(). This caused a NULL pointer dereference
issue in cases where a rule could not be found in the tree.

Signed-off-by: Awais Belal <awais.belal@mentor.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>